build: Update `@types/node` after TypeScript 5.9 upgrade
With TypeScript 5.9 there were errors in the outdated node types that
we intentionally used to ensure compatibility with our minimum supported
Node.js version. In order to solve them, we need to upgrade the types to
the first version which corrected them, which is 20.15.

Our CI still ensures compatibility with Node 20.11.0.