fix(rules/renderer): Detect missing apiVersion for method declarations
Similar to the already supported property assigment with an arrow
function, 'renderer' might also be declared as a method declarations or
function declaration (which is transpiled to a method declaration).

If the 'renderer' property is set to a function/method, this implies
apiVersion 1 and should therefore be reported as deprecated API usage.

Follow up of https://github.com/UI5/linter/pull/439

JIRA: CPOUI5FOUNDATION-878