ci: Fail-fast "npm ci" if deps require unsupported node version (#912)
When upgrading a dependency to a newer version, it might require a
different node version than before which could be unsupported by the
repo. Unfortunately, this could be easily overseen.
This PR adds the flag `--engine-strict` to the dependency installation
step during the GH Actions job. Incase such a version requirement
change, this flag will now fail-fast the job which should attract
attention.